+++ src/data/articlesStore.js
@@ -3,22 +3,22 @@ const DEFAULT_ARTICLES = [
     {
         id: 1,
         title: "Bodhichitta",
-        creationDate: "2024-11-24",
-        lastEditDate: "2025-01-10",
+        creationDate: "2025-03-02T12:55:00",
+        lastEditDate: "2025-03-02T12:55:00",
         content: "Need to check on human, have not seen in an hour might be dead oh look, human is alive, hiss at human, feed me. Warm up laptop with butt lick butt fart rainbows until owner yells pee in litter box hiss at cats poop on the floor, break a planter, sprint, eat own hair, vomit hair, hiss, chirp at birds, eat a squirrel, hide from fireworks, lick toe beans, attack christmas tree but show belly."
     },
     {
         id: 2,
         title: "Smile",
-        creationDate: "2024-10-28",
-        lastEditDate: "2024-11-02",
+        creationDate: "2025-03-01T12:55:00",
+        lastEditDate: "2025-03-01T12:55:00",
         content: "Need to check on human, have not seen in an hour might be dead oh look, human is alive, hiss at human, feed me. Warm up laptop with butt lick butt fart rainbows until owner yells pee in litter box hiss at cats poop on the floor, break a planter, sprint, eat own hair, vomit hair, hiss, chirp at birds, eat a squirrel, hide from fireworks, lick toe beans, attack christmas tree but show belly."
     },
     {
         id: 3,
         title: "The walking meditation",
-        creationDate: "2024-08-24",
-        lastEditDate: "2024-08-24",
+        creationDate: "2023-03-01T12:55:00",
+        lastEditDate: "2023-03-01T12:55:00",
         content: "Need to check on human, have not seen in an hour might be dead oh look, human is alive, hiss at human, feed me. Warm up laptop with butt lick butt fart rainbows until owner yells pee in litter box hiss at cats poop on the floor, break a planter, sprint, eat own hair, vomit hair, hiss, chirp at birds, eat a squirrel, hide from fireworks, lick toe beans, attack christmas tree but show belly."
     }
 ];